# artifacts/api-server/src/routes/export.ts
import { Router, type IRouter } from "express";
import { eq } from "drizzle-orm";
import { db, projectsTable, entitiesTable, propertiesTable, rulesTable } from "@workspace/db";
import { sql } from "drizzle-orm";

const router: IRouter = Router();

router.get("/projects/:projectId/export/tabletop-simulator", async (req, res): Promise<void> => {
  const projectId = parseInt(Array.isArray(req.params.projectId) ? req.params.projectId[0] : req.params.projectId, 10);
  if (isNaN(projectId)) {
    res.status(400).json({ error: "Invalid project ID" });
    return;
  }

  const [project] = await db.select().from(projectsTable).where(eq(projectsTable.id, projectId));
  if (!project) {
    res.status(404).json({ error: "Project not found" });
    return;
  }

  const entities = await db.select().from(entitiesTable).where(eq(entitiesTable.projectId, projectId));
  const entityIds = entities.map((e) => e.id);
  let propertiesByEntity: Record<number, typeof propertiesTable.$inferSelect[]> = {};
  if (entityIds.length > 0) {
    const allProps = await db.select().from(propertiesTable).where(
      sql`${propertiesTable.entityId} = ANY(${sql.raw(`ARRAY[${entityIds.join(",")}]`)})`
    );
    for (const prop of allProps) {
      if (!propertiesByEntity[prop.entityId]) propertiesByEntity[prop.entityId] = [];
      propertiesByEntity[prop.entityId].push(prop);
    }
  }

  // Build Tabletop Simulator compatible object states
  const objectStates = entities.map((entity, idx) => {
    const props = propertiesByEntity[entity.id] ?? [];
    const customFields: Record<string, string | number> = {};
    for (const prop of props) {
      customFields[prop.name] = prop.defaultValue ?? (prop.dataType === "number" ? 0 : "");
    }

    return {
      GUID: `${projectId}${entity.id}${idx}`.padStart(6, "0"),
      Name: "Card",
      Nickname: entity.name,
      Description: entity.description ?? "",
      Tags: [entity.type],
      CustomDeck: {
        "1": {
          FaceURL: "",
          BackURL: "",
          NumWidth: 1,
          NumHeight: 1,
          BackIsHidden: false,
          UniqueBack: false,
        },
      },
      CardID: idx + 1,
      CustomFields: customFields,
      Transform: {
        posX: idx * 2.5,
        posY: 1,
        posZ: 0,
        rotX: 0,
        rotY: 0,
        rotZ: 0,
        scaleX: 1,
        scaleY: 1,
        scaleZ: 1,
      },
    };
  });

  res.json({
    objectStates,
    savedAt: new Date().toISOString(),
  });
});

router.get("/projects/:projectId/export/rulebook", async (req, res): Promise<void> => {
  const projectId = parseInt(Array.isArray(req.params.projectId) ? req.params.projectId[0] : req.params.projectId, 10);
  if (isNaN(projectId)) {
    res.status(400).json({ error: "Invalid project ID" });
    return;
  }

  const [project] = await db.select().from(projectsTable).where(eq(projectsTable.id, projectId));
  if (!project) {
    res.status(404).json({ error: "Project not found" });
    return;
  }

  const entities = await db.select().from(entitiesTable).where(eq(entitiesTable.projectId, projectId));
  const entityIds = entities.map((e) => e.id);
  let propertiesByEntity: Record<number, typeof propertiesTable.$inferSelect[]> = {};
  if (entityIds.length > 0) {
    const allProps = await db.select().from(propertiesTable).where(
      sql`${propertiesTable.entityId} = ANY(${sql.raw(`ARRAY[${entityIds.join(",")}]`)})`
    );
    for (const prop of allProps) {
      if (!propertiesByEntity[prop.entityId]) propertiesByEntity[prop.entityId] = [];
      propertiesByEntity[prop.entityId].push(prop);
    }
  }

  const rules = await db.select().from(rulesTable).where(eq(rulesTable.projectId, projectId)).orderBy(rulesTable.priority);

  const lines: string[] = [
    `# ${project.name} — Rulebook`,
    ``,
    `*Generated by AI Board Game Factory*`,
    ``,
    `---`,
    ``,
  ];

  if (project.description) {
    lines.push(`## About`, ``, project.description, ``);
  }

  const meta: string[] = [];
  if (project.genre) meta.push(`**Genre:** ${project.genre}`);
  if (project.playerCount) meta.push(`**Players:** ${project.playerCount}`);
  if (project.targetDuration) meta.push(`**Duration:** ${project.targetDuration}`);
  if (meta.length > 0) {
    lines.push(...meta, ``);
  }

  if (entities.length > 0) {
    lines.push(`## Components`, ``, `### Game Entities`, ``);
    for (const entity of entities) {
      const props = propertiesByEntity[entity.id] ?? [];
      lines.push(`#### ${entity.name} *(${entity.type})*`);
      if (entity.description) lines.push(``, entity.description);
      if (props.length > 0) {
        lines.push(``, `**Properties:**`, ``);
        for (const prop of props) {
          let propDesc = `- **${prop.name}** (${prop.dataType})`;
          if (prop.defaultValue) propDesc += ` — default: ${prop.defaultValue}`;
          if (prop.minValue || prop.maxValue) propDesc += ` [${prop.minValue ?? ""}–${prop.maxValue ?? ""}]`;
          if (prop.unit) propDesc += ` ${prop.unit}`;
          if (prop.description) propDesc += `. ${prop.description}`;
          lines.push(propDesc);
        }
      }
      lines.push(``);
    }
  }

  if (rules.length > 0) {
    const byCategory: Record<string, typeof rules> = {};
    for (const rule of rules) {
      const cat = rule.category ?? "General";
      if (!byCategory[cat]) byCategory[cat] = [];
      byCategory[cat].push(rule);
    }

    lines.push(`## Rules`, ``);
    for (const [cat, catRules] of Object.entries(byCategory)) {
      lines.push(`### ${cat.charAt(0).toUpperCase() + cat.slice(1)}`, ``);
      for (const rule of catRules) {
        lines.push(`#### ${rule.title}`, ``, rule.content, ``);
      }
    }
  }

  lines.push(`---`, ``, `*Rulebook generated ${new Date().toLocaleDateString()}*`);

  res.json({
    markdown: lines.join("\n"),
    title: project.name,
    generatedAt: new Date().toISOString(),
  });
});

export default router;