# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Board Game Factory API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/entities:
    get:
      operationId: listEntities
      tags: [entities]
      summary: List all entities in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of entities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entity"
    post:
      operationId: createEntity
      tags: [entities]
      summary: Create a new entity in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntityBody"
      responses:
        "201":
          description: Created entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
  /projects/{projectId}/entities/{id}:
    get:
      operationId: getEntity
      tags: [entities]
      summary: Get an entity with its properties
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entity with properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityWithProperties"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEntity
      tags: [entities]
      summary: Update an entity
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntityBody"
      responses:
        "200":
          description: Updated entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEntity
      tags: [entities]
      summary: Delete an entity
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Properties ────────────────────────────────────────────────────────────
...
  /projects/{projectId}/assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List all generated assets for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create/save an asset
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetBody"
      responses:
        "201":
          description: Created asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /projects/{projectId}/assets/{id}:
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete an asset
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/generate-card-description:
    post:
      operationId: generateCardDescription
      tags: [assets]
      summary: Generate card flavor text and description using AI (SSE stream)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCardDescriptionBody"
      responses:
        "200":
          description: SSE stream of generated description
          content:
            text/event-stream: {}
  /projects/{projectId}/generate-asset-image:
    post:
      operationId: generateAssetImage
      tags: [assets]
      summary: Generate an image for a game asset using AI
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAssetImageBody"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedImageResponse"

  # ── Export ────────────────────────────────────────────────────────────────
  /projects/{projectId}/export/tabletop-simulator:
    get:
      operationId: exportTabletopSimulator
      tags: [export]
      summary: Export project as Tabletop Simulator JSON object states
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tabletop Simulator export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TabletopSimulatorExport"
  /projects/{projectId}/export/rulebook:
    get:
      operationId: exportRulebook
      tags: [export]
      summary: Export project as a Markdown rulebook
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Markdown rulebook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RulebookExport"

  # ── Anthropic (Rules Sandbox AI) ──────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Projects ──────────────────────────────────────────────────────────
...
    Entity:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        type:
          type: string
          description: "e.g. Item, Faction, Location, Event"
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - type
        - createdAt
        - updatedAt
    EntityWithProperties:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        type:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        properties:
          type: array
          items:
            $ref: "#/components/schemas/Property"
      required:
        - id
        - projectId
        - name
        - type
        - createdAt
        - updatedAt
        - properties
    CreateEntityBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
      required:
        - name
        - type
    UpdateEntityBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]

    # ── Properties ────────────────────────────────────────────────────────
    Property:
      type: object
      properties:
        id:
          type: integer
        entityId:
          type: integer
        name:
          type: string
        dataType:
          type: string
          description: "e.g. number, string, boolean, enum"
        defaultValue:
          type: ["string", "null"]
        minValue:
          type: ["string", "null"]
        maxValue:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - entityId
        - name
        - dataType
        - createdAt
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        assetType:
          type: string
          description: "e.g. card, token, board_tile"
        entityId:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        flavorText:
          type: ["string", "null"]
        imageData:
          type: ["string", "null"]
          description: Base64 encoded image or URL
        imagePrompt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - assetType
        - createdAt
    CreateAssetBody:
      type: object
      properties:
        name:
          type: string
        assetType:
          type: string
        entityId:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        flavorText:
          type: ["string", "null"]
        imageData:
          type: ["string", "null"]
        imagePrompt:
          type: ["string", "null"]
      required:
        - name
        - assetType
    GenerateCardDescriptionBody:
      type: object
      properties:
        entityId:
          type: ["integer", "null"]
        cardName:
          type: string
        cardType:
          type: string
        context:
          type: ["string", "null"]
      required:
        - cardName
        - cardType
    GenerateAssetImageBody:
      type: object
      properties:
        prompt:
          type: string
        style:
          type: ["string", "null"]
          description: "e.g. fantasy illustration, pixel art, watercolor"
        size:
          type: string
          enum: ["1024x1024", "1536x1024", "1024x1536"]
      required:
        - prompt
    GeneratedImageResponse:
      type: object
      properties:
        b64_json:
          type: string
      required:
        - b64_json

    # ── Export ────────────────────────────────────────────────────────────
    TabletopSimulatorExport:
      type: object
      properties:
        objectStates:
          type: array
          items:
            type: object
        savedAt:
          type: string
          format: date-time
      required:
        - objectStates
        - savedAt
    RulebookExport:
      type: object
      properties:
        markdown:
          type: string
        title:
          type: string
        generatedAt:
          type: string
          format: date-time
      required:
        - markdown
        - title
        - generatedAt

    # ── Anthropic ─────────────────────────────────────────────────────────
...