# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Board Game Factory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Asset {
  id: number;
  projectId: number;
  name: string;
  /** e.g. card, token, board_tile */
  assetType: string;
  /** @nullable */
  entityId?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  flavorText?: string | null;
  /**
   * Base64 encoded image or URL
   * @nullable
   */
  imageData?: string | null;
  /** @nullable */
  imagePrompt?: string | null;
  createdAt: Date;
}