# lib/api-zod/src/generated/types/generateCardDescriptionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Board Game Factory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateCardDescriptionBody {
  /** @nullable */
  entityId?: number | null;
  cardName: string;
  cardType: string;
  /** @nullable */
  context?: string | null;
}