# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Board Game Factory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: number;
  entityId: number;
  name: string;
  /** e.g. number, string, boolean, enum */
  dataType: string;
  /** @nullable */
  defaultValue?: string | null;
  /** @nullable */
  minValue?: string | null;
  /** @nullable */
  maxValue?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt: Date;
}