# lib/api-zod/src/generated/types/tabletopSimulatorExport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Board Game Factory API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TabletopSimulatorExportObjectStatesItem } from "./tabletopSimulatorExportObjectStatesItem";

export interface TabletopSimulatorExport {
  objectStates: TabletopSimulatorExportObjectStatesItem[];
  savedAt: Date;
}